remediate REV-0001 + REV-0002 independent-review findings (WO-0013/0014/0015)
Human-authorized remediation of the confirmed cross-model review findings
(Codex/GPT-5 in work/review/REV-0001/result.md + REV-0002/result.md). All three
changes are additive on human-gated surfaces; approach + design decisions were
approved before implementation.

WO-0013 — complete the event-truth flip on the order-status WRITE path (REV-0001):
- F-001 (P0): claim_order_for_submission now derives status from the event-log
  projection under the same lock (both stores), not the raw orders.status column,
  so a drifted column can no longer re-claim an already-submitted order. Safe
  because the co-write invariant holds (every orders.status write co-appends a
  lifecycle event -> projection is complete; verified by auditing every writer).
- F-002 (P0): the init backfill keys on status-lifecycle-event ABSENCE (new
  projectors.ORDER_STATUS_EVENT_TYPES, which excludes FILL), not projected==CREATED
  -- a released order (SUBMIT_PENDING->SUBMIT_RELEASED, projects CREATED but has
  events) is no longer clobbered by a synthetic terminal event; a pre-eventing
  FILLED order (fills, no lifecycle event) is still reconstructed.
- F-003: ADR-008 amended to cover SUBMIT_RELEASED / CANCEL_PENDING provenance
  (ENGINE/LOCAL) with the ADR-001 consistency note.
- Sibling write paths (transition_order/flatten/reconcile) documented safe under
  the verified co-write invariant rather than rewritten (a full projection flip of
  every write path is a possible future WO).
- 10 dual-store RED->GREEN regressions (tests/test_wo0013_event_truth_writepath.py).

WO-0014 — market-data fractional volume (REV-0002 F-003):
- volume int->float end-to-end (MarketSnapshot, MarketSnapshotView, fake, seed
  annotations); removed the int() that truncated fractional trade.size before the
  min-volume gate. 2 RED->GREEN regressions.

WO-0015 — flatten deferred-visibility + actor provenance (REV-0002 F-001/F-002):
- F-001: a manual flatten safely DEFERRED to a live protection exit is now an
  explicit operator-visible state -- FlattenResult.deferred / FlattenResponse
  .deferred + deferred_order_status (keyed on the deferral event, so an idempotent
  own-flatten stays deferred=False), and the cockpit renders a distinct "no manual
  order submitted -- already exiting; monitoring" message instead of "submitted".
- F-002: the command actor is threaded route->facade->store and recorded on both
  the manual_flatten_deferred and the created manual-flatten's sell_intent_created
  event payloads (default "system"; protection-tick create stays "system").
- Purely additive: no change to the flatten decision, no-blind-cancel behavior, or
  any order/fill/position state (INV-9). INVARIANTS INV-034/036 follow-ups resolved.
- Dual-store + facade + cockpit RED->GREEN tests.

Verified on the combined tree: full suite 1981 passed / 5 skipped / 0 failed;
ruff check clean; mypy app/ Success (54 files); import-linter 5 kept / 0 broken.

+++ app/marketdata/alpaca_stream.py
@@ -109,7 +109,7 @@ def _trading_day(dt: datetime) -> date:
 
 
 def _seed_from_snapshot(raw) -> tuple[
-    Optional[float], Optional[float], Optional[float], Optional[int], Optional[float]
+    Optional[float], Optional[float], Optional[float], Optional[float], Optional[float]
 ]:
     """Extract ``(last_price, bid, ask, volume, prev_close)`` from an Alpaca
     ``Snapshot``. Any missing sub-object yields ``None`` for its fields — a
@@ -316,7 +316,7 @@ class AlpacaMarketDataStream(MarketDataService):
 
     def _fetch_seed(
         self, symbol: str
-    ) -> tuple[Optional[float], Optional[float], Optional[float], Optional[int], Optional[float]]:
+    ) -> tuple[Optional[float], Optional[float], Optional[float], Optional[float], Optional[float]]:
         """REST-seed one symbol's snapshot fields (sync; called via to_thread).
 
         One symbol at a time, not a batched multi-symbol request: a batch
@@ -413,9 +413,11 @@ class AlpacaMarketDataStream(MarketDataService):
                 # trades that occurred during the gap) — adequate for a
                 # threshold gate (Strategy Engine's min-volume check), not for
                 # precise reporting.
-                # volume is a whole-share count (MarketSnapshot.volume is int);
-                # trade.size is typed float by the SDK, so coerce the sum back to int.
-                volume=int((existing.volume or 0) + trade.size),
+                # trade.size is typed float by the SDK (fractional / odd-lot prints
+                # occur), and MarketSnapshot.volume is float, so keep the running sum
+                # as a float — an int() coercion here silently truncated sub-share
+                # sizes before the min-volume gate (REV-0002 F-003).
+                volume=(existing.volume or 0) + trade.size,
                 updated_at=utcnow(),
             )
